+++ example/lib/generated_plugin_registrant.dart
@@ -2,9 +2,10 @@
 // Generated file. Do not edit.
 //
 
+// ignore_for_file: directives_ordering
 // ignore_for_file: lines_longer_than_80_chars
 
-import 'package:file_picker/src/file_picker_web.dart';
+import 'package:file_picker/_internal/file_picker_web.dart';
 import 'package:shared_preferences_web/shared_preferences_web.dart';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';